Time: 3 ms (37.90%), Space: 5.8 MB (95.11%) - LeetHub

+++ 0070-climbing-stairs/0070-climbing-stairs.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int climbStairs(int n) {
         const int MaxTwoCount = n / 2;
-        long long Fraction = n - 1;
+        long Fraction = n - 1;
         int Ways = n;
         
         for (int i = 2; i <= MaxTwoCount; i++)
